[front] chore: Migrate Databrick MCP
